Added GPSDataPopup.java in the GUI package. It creates a popup where the user selects the test case. A new GPSDataTransfer object is then created, but I don't really know what to do with it.
Added a second parameter to GPSDataTransfer.java, a string that varies depending on which test case is selected in the popup.

Commented out the old GPS Data Transfer code in MarsHabitatConfigurationSystem.java and tried to add the new popup. I just followed what was done for the Add Module Popup.


NOTE: Some errors were in the project when I checked it out to work on it so I am unable to compile and test my changes. Everything SHOULD work in theory, I just need to ask you guys what to do with the created GPS object in the GPSDataPopup.java